patch de atualização

+++ commands/removerbemvindo.js
@@ -11,9 +11,34 @@ module.exports = {
 
     //quick.db
 
+    if (user.id === '380493625227870222') {
+      db.delete(`welchannel_${message.guild.id}`)
+      db.delete(`welmessage_${message.guild.id}`)
+      const embed = new Discord.MessageEmbed()
+      .setAuthor('Mensagem de aviso')
+      .setDescription('Mensagem de boas vindas foi removida.')
+      .setColor('#cc0000')
+      .setFooter(`requirido por • ${message.author.tag}`, message.author.displayAvatarURL({ format: "png" }))
+      return message.channel.send(embed).then(msg => {
+        msg.delete({ timeout: 5000 })
+      })
+        .catch(console.error);
+    }
+
+    if(!message.member.permissions.has("MANAGE_MESSAGES")) {
+      embed.setAuthor('Mensagem de erro')
+      embed.setDescription(`${message.author.username}, você não tem permissão para usar este comando!`)
+      embed.setFooter(`requirido por • ${message.author.tag}`, message.author.displayAvatarURL({ format: "png" }))
+      embed.setColor('#cc0000')
+      return message.channel.send(embed).then(msg => {
+        msg.delete({ timeout: 10000 })
+      })
+        .catch(console.error);
+    }
+
     db.delete(`welchannel_${message.guild.id}`)
     db.delete(`welmessage_${message.guild.id}`)
-    const embed = new discord.MessageEmbed()
+    const embed = new Discord.MessageEmbed()
     .setAuthor('Mensagem de aviso')
     .setDescription('Mensagem de boas vindas foi removida.')
     .setColor('#cc0000')
